move extraction to dedicated submodule
