+++ readabilipy/extractors/__init__.py
@@ -1,0 +1,3 @@
+from .extract_title import extract_title
+
+__all__ = ['extract_title']
