+++ readabilipy/json_parser.py
@@ -6,7 +6,7 @@ from subprocess import check_call
 from bs4 import BeautifulSoup
 from bs4.element import Comment, NavigableString, CData
 from .plain_html import parse_to_tree
-from .plain_html import extract_title
+from .extractors.extract_title import extract_title
 from .text_manipulation import normalise_text
